requirements.txt
altgraph==0.17.3
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
numpy==1.25.2
oauth2==1.9.0.post1
pefile==2023.2.7
psycopg2==2.9.6
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.31.0
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
